+++ CHANGELOG.md
@@ -8,6 +8,8 @@ FEATURES:
 
 IMPROVEMENTS:
 
+* node: Limit number of concurrent gossip routines
+
 BUG FIXES:
 
 ## v0.5.3 (August 29, 2019)