+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 IMPROVEMENTS:
 
 - badger_db: Update to use latest v1.6.0 release 
+- config: Enable setting a different database directory with --db
 
 
 ## v0.5.7 (October 2, 2019)